perceiver is functional. Ready to unit test the monster
