+++ pytorch_widedeep/training/trainer.py
@@ -24,9 +24,9 @@ from pytorch_widedeep.callbacks import (
 from pytorch_widedeep.dataloaders import DataLoaderDefault
 from pytorch_widedeep.initializers import Initializer, MultipleInitializer
 from pytorch_widedeep.training._finetune import FineTune
+from pytorch_widedeep.utils.general_utils import Alias
 from pytorch_widedeep.training._wd_dataset import WideDeepDataset
 from pytorch_widedeep.training.trainer_utils import (
-    Alias,
     alias_to_loss,
     save_epoch_logs,
     wd_train_val_split,
